Rename /web/public to /web/www
If we ever plan to wrap the app around Cordova in the future, the name
of the distribution directory must be www (AFAIK, Cordova does not allow
you to configure this). The directory is being renamed now so that the
front-end developers will not need to change their virtual host
configuration in the future.